feat: 168 (refs to persist data between renders), 169 (custom hooks), 170 (useMovies custom hook), 171 (useLocalStorageState custom hook), 172 (useKey custom hook)
