+++ 07-usepopcorn/src/useMovies.js
@@ -1,0 +1,61 @@
+import { useState, useEffect } from "react";
+
+
+const API_KEY = "8c99f3d0";
+
+
+// we cannot use the fetch and setState inside the component as it is causing
+// infinite re-rendering - that's why useEffect
+export function useMovies(query) {
+    const [movies, setMovies] = useState([]);
+    const [isLoading, setIsLoading] = useState(false);
+    const [error, setError] = useState("");
+
+    useEffect(function () {
+        // to close the selected movie when new search starts
+        // callback?.();
+
+        const controller = new AbortController();
+
+        async function fetchMovies() {
+            try {
+                setIsLoading(true);
+                setError("");
+                const res = await fetch(
+                    `http://www.omdbapi.com/?apikey=${API_KEY}&s=${query}`,
+                    {signal: controller.signal}
+                );
+                if(!res.ok)
+                    throw new Error("Something went wrong ...");
+                const data = await res.json();
+                if (data.Response === "False")
+                    throw new Error("Movie not found!")
+                setMovies(data.Search);
+                setError("");
+            } catch (err) {
+                if (err.name !== "AbortError") {
+                    setError(err.message);
+                }
+            } finally {
+                setIsLoading(false);
+            }
+        }
+
+        if(query.length < 3) {
+            setMovies([]);
+            setError("");
+            return;
+        }
+
+        fetchMovies();
+
+        return function() {
+            controller.abort();
+        };
+
+        }, [query]
+        );
+
+        return {movies, isLoading, error}
+
+}
